fix(authController) made a couple of fixes to authController functions and updated authSwagger accordingly
