+++ backend/src/jest.setup.ts
@@ -1,7 +1,7 @@
 import dotenv from 'dotenv';
 
 // Load test environment variables
-dotenv.config({ path: '.env.test' });
+dotenv.config({ path: '.env' }); // use .env.test if needed
 
 // Mock console errors to keep test output clean
 jest.spyOn(console, 'error').mockImplementation(() => {});
