+++ backend/src/server.ts
@@ -6,8 +6,6 @@ const MONGO_URI = process.env.MONGO_URI;
 
 mongoose
   .connect(MONGO_URI, {
-    useNewUrlParser: true,
-    useUnifiedTopology: true,
   } as mongoose.ConnectOptions)
   .then(() => {
     console.log('MongoDB connected successfully');
